pyproject.toml
[project]
name = "tripadvisor-foodbot"
version = "0.1.0"
description = "TripAdvisor Food Bot"
readme = "README.md"
requires-python = "~=3.11"
dependencies = [
    "asyncpg>=0.30.0",
    "boto3==1.38.26",
    "chainlit==2.5.5",
    "db-dtypes==1.4.2",
    "fastembed==0.6.1",
    "google-cloud-bigquery==3.31.0",
    "google-cloud-bigquery-storage==2.30.0",
    "llama-index==0.12.30",
    "loguru==0.7.3",
    "numba==0.61.2",
    "pandas==2.2.3",
    "pandas-gbq>=0.29.1",
    "pytest==8.3.5",
    "qdrant-client==1.13.3",
    "scipy>=1.15.3",
    "sqlparse==0.5.3",
    "tabulate==0.9.0",
    "unidecode==1.4.0",
    "uuid==1.30",
    "watchdog==6.0.0",
]

[dependency-groups]
dev = [
    "ruff==0.11.5",
]

[tool.ruff]
line-length = 150
exclude = [".venv", "build", "dist", "include"]

[tool.ruff.lint]
select = ["F", "I001", "RUF022"]

[tool.uv]
package = true
